Fixed extended teamwork bug, more correct teamwork results.
The new teamwork method was only adding a maximum of one teammate
per roll, which was a problem when multiple teammates had equal
size dice pools.  Fixed.

Changed extended teamwork glitches to increase threshold, instead
of reducing hits.  The effect is the same, but this way is
technically correct, and it reports a bit nicer.
That is, getting negative hits on a test was a bit confusing.