Standard test now a single iteration extended test.
Code reuse is good.
Removed standard().  Instead, use @standard to mark a
single iteration extended test for use in extended().

Changed teamwork extended glitches to be hits -= 3, instead of
threshold += 1.  This is the literal wording in the book now.

Added a bit of pretty printing for teamwork tests.
Each iteration now appears as it's own text block in the output.

Moved some --help items around to group them.
In particular, tried to get the extended test modifiers together.